refactor: merge auth/config.ts, auth/auth.ts
