+++ global.d.ts
@@ -1,4 +1,4 @@
-import { User as ZiggleUser } from '@/api/auth/config';
+import { User as ZiggleUser } from '@/api/auth/auth';
 
 declare module '*.svg' {
   const content: React.FC<React.SVGProps<SVGSVGElement>>;
