+++ src/app/components/templates/Sidebar/SidebarMobile.tsx
@@ -5,7 +5,7 @@ import Link from 'next/link';
 import { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 
-import { User } from '@/api/auth/config';
+import { User } from '@/api/auth/auth';
 import { PropsWithLng } from '@/app/i18next';
 import { useTranslation } from '@/app/i18next/client';
 import DefaultProfile from '@/assets/default-profile.svg';
